Unit9/src/security/authentication.py
from abc import ABC, abstractmethod
from typing import Optional
from dataclasses import dataclass
import jwt

@dataclass
class AuthToken:
    user_id: int
    token: str
    expires_at: int

class AuthenticationService:
    def __init__(self, user_service, secret_key: str):
        self._user_service = user_service
        self._secret_key = secret_key
    
    def authenticate(self, username: str, password: str) -> Optional[AuthToken]:
        user = self._user_service.find_by_username(username)
        if user and self._verify_password(password, user._password):
            return self._create_token(user.id)
        return None
    
    def _create_token(self, user_id: int) -> AuthToken:
        # Implementation of JWT token creation
        pass