Unit9/requirements.txt
# Core dependencies
pyjwt>=2.4.0
bcrypt>=4.0.1
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# API dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.4.0
# Removed types-jwt as it's not needed with recent pyjwt versions

# Development tools
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0